WIP: refactoring replay code into its own set of files
